feat: splits the button component into three subcomponents to reduce code
